requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
numpy
opencv-python
ultralytics
easyocr
